refactor(auth): move persistant login logic to a custom hook - add useLoadInitialAuthState hook
