+++ src/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './useLogout';
+export * from './useLoadInitialAuthState';
